+++ packages/component-library/src/style/breakpoints.ts
@@ -1,9 +1,10 @@
-export const breakpointNames = ['mobile', 'tablet', 'desktop'] as const;
-
 export const breakpoints = {
   mobile: 0,
-  tablet: 50,
-  desktop: 62,
+  tablet: 740,
+  desktop: 992,
+  wide: 1200,
 } as const;
 
 export type Breakpoint = keyof typeof breakpoints;
+
+export const breakpointNames = ['mobile', 'tablet', 'desktop', 'wide'] as const;