+++ packages/component-library/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cutting/component-library
 
+## 5.3.1
+
+### Patch Changes
+
+- conver vars.space to rem
+
 ## 5.3.0
 
 ### Minor Changes